requirements.txt
Faker==37.4.0
lxml==5.4.0
numpy==2.3.0
pandas==2.3.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
six==1.17.0
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
